app/requirements.txt
django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
pytest-django==4.8.0
pytest==8.2.2
celery[redis]==5.4.0
django-redis==5.4.0
django-cachalot==2.6.2
requests==2.32.3
beautifulsoup4==4.12.3
pytest-cov==5.0.0
flake8==7.1.0
black==24.4.2
isort==5.13.2
gunicorn==22.0.0
django-cors-headers==4.4.0
tradingview-ta